+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add comment to shared note/link
 * Add copy buttons for restoring archived projects
 * Fix Bitwarden auto-fill on login form
+* Fix spellcheck disabled when settings are not loaded yet
 
 
 ## v2025.90 - 2025-10-22